+++ src/components/Terminal/utils/collapseLine.js
@@ -1,8 +1,0 @@
-const collapseLine = (line) => {
-  return line
-    .filter((token) => !token.types.includes('comment'))
-    .map((token) => token.content)
-    .join('');
-};
-
-export default collapseLine;