+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+# [1.32.0](https://github.com/newrelic/developer-website/compare/v1.31.8...v1.32.0) (2021-01-15)
+
+
+### Bug Fixes
+
+* upgrade gatsby, gatsby-remark-autolink-headers ([29c47cf](https://github.com/newrelic/developer-website/commit/29c47cfa3dc02a218bb166cf13f155ffa5356612))
+
+
+### Features
+
+* add feedback component to right rail ([7e19c02](https://github.com/newrelic/developer-website/commit/7e19c02fcd0516089e31d9e2960639ecfca2e6a7))
+* add icon for headers to make it easy to link to them ([a254146](https://github.com/newrelic/developer-website/commit/a254146b04a9b9e4cb4dab2262661189ba327485))
+
+## [1.31.8](https://github.com/newrelic/developer-website/compare/v1.31.7...v1.31.8) (2021-01-15)
+
+
+### Bug Fixes
+
+* upgrade @xstate/react from 1.0.3 to 1.1.0 ([cc9bd5d](https://github.com/newrelic/developer-website/commit/cc9bd5d0a762cfa3e734210500d76e822d720d6c))
+
 ## [1.31.7](https://github.com/newrelic/developer-website/compare/v1.31.6...v1.31.7) (2020-12-18)